requirements.txt
fastapi==0.109.2
pymongo==4.6.1
pydantic==2.6.1
python-dotenv==1.0.1
pydantic-settings==2.2.0
google-api-python-client==2.118.0
python-dateutil==2.8.2
uvicorn==0.27.1